apps/01-hippoRAG-evaluation/requirements.txt
# Core dependencies
flashrag>=0.1.0
openai>=1.12.0
tiktoken>=0.5.0

# Vector stores
faiss-cpu>=1.7.4
numpy>=1.24.0

# Sparse retrieval
rank-bm25>=0.2.2

# NLP and entity recognition
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Graph processing
networkx>=3.1

# Data processing
datasets>=2.16.0
pandas>=2.0.0
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
